glofiish: Add a simple state machine for tracking the modem communication
With this simple state machine it should be possible to trace the various
configurations steps knight does to speak with the modem. Currently there
is some weird behaiviour, that the state variable is changing to a random
value and I don't know why. It is only set to the initial state and then
never touch again, but when the first state change comes, it has a very
big random value.